Fix Bug 226 - querystring widget bugs
This commit addresses the following issues reported in #226:

* "remove line" shouldn't be in line where to select new criteria.
Added a css rule to not show the x on the last line.
* "remove line" should be an icon
An icon (css content char) was already there. Added css to reduce width
of field.
* input fields should be as high as dropdowns (e.g. date selection in
* querystring).
Added css statements to remove default margin so that the lines below
are not pushed down when date field inputs are inserted. Also added an
extra clear element to separate criteria from sort section.
* creator selection "is" doesn't work - no autosuggestion of existing
* users.
Added an UserPickerWidget to allow autosuggest for users, added a new
json test data provider in fakeserver.js
* location/"absolute path" (maybe also "relative path") should show
* related items widget
Added the RelatedItems pattern for the absolute path location. The
relative path should stay a string widget according to Rok.
* date selection "before/after date", between dates - datetime selection
* instead of date
Removed the time: false configuration for the date pickers.

* also created tests for querystring