Fire "focusin" event instead of "focus".
This fixes failing tests where the event handler for "focus" wasn't being
called.

From https://api.jquery.com/on :

"The focus and blur events are specified by the W3C to not bubble, but jQuery
defines cross-browser focusin and focusout events that do bubble. When focus
and blur are used to attach delegated event handlers, jQuery maps the names and
delivers them as focusin and focusout respectively. For consistency and
clarity, use the bubbling event type names."