Bugfix. See below.
For reference, see my comment here:
https://github.com/Patternslib/Patterns/issues/396#issuecomment-73722048

If we have multiple bundles that all pull in pat-registry and then call
registry.scan, then patterns will be scanned and initialized multiple times.

Instead, we need to call registry.init *if* it's not been called before. If
it's been called before, we don't need to do anything because the patterns will
automatically be registered as soon as they are loaded by require.js and evaluated
by the JS engine and then automatically scanned by the registry.