Revert "fix autotoc does not change to internal link after upload"
This reverts commit aa103d46d5d8760c22b5788336020dca131dcd50.

After pull request #602 was merged, Jenkins gave an error in
CMFEditions (test15_storageStatistics) in the AT tests:
http://jenkins.plone.org/job/plone-5.0-python-2.7-at/3376/
The error is persistent.  First thing to try is to revert these commits.

Note that I don't see a link between the error and these commits,
though maybe the tinymce changes in here influence what gets saved in
CMFEditions.

And note that I do not get the test failure locally.  Joy...