Fix manifest to not include node_modules, even when later removed.
The coredev buildout took very long for me.
Trying it manually, `python setup.py egg_info` took five minutes.
Problem was that I had a large `mockup/node_modules` directory.
This was pruned in `MANIFEST.in`, so it never ended up in the release.
But is *was* first included before being pruned, and this took extremely long.

[ci skip]