maint(Bundle): Move dependency backbone.paginator to devDependencies to avoid version conflicts.
backbone.paginator is now in devDependencies as it requests outdated
versions of Backbone and Underscore. Having this in normal dependencies
field would eventually let backbone.paginator register itself in an
outdated Backbone version and make it unavailable by those used in
structure pattern.
Having this in devDependencies avoids this, but also makes the structure
pattern unavailable in add-on bundles unless this dependency is
explicitly added.
However, this should be fixed in a more sane way. But that's out of
scope for now.