fix(pat-datatables): Fix tests.
Fix pat-structure test by requiring datatables from the test-setup
beforehand loading the test files. This is a necessary step for
datatables which wouldn't otherwise be attached to jQuery in tests.
Also freshly wrap the element with jQuery in the Patterns init method
after datatables was imported. A non-jQuery setup does not work at the
moment.