fix(pat autotoc): Show required indicator for tabs with reuqired multi-checkbox or radio button fields.
A groupt of multiple required checkboxes for a single field are rendered
in Plone currently without the required attribute. The same goes for radio
buttons. Instead the label is marked as required.

This PR supports this situation and also marks the tab-navigation as
required, if a field label is marked as required.`