fix(pat-tinymce): Add a fallback for Plone 5 images without data-picturevariant attributes.
Responsive picture variants were introduced in Plone 6. Content from
Plone 5 where no `data-picturevariant` was available lost had no scale
selected in the image dialog.
This PR fixes that by falling back to the `data-scale` attribute.
This allows old images in the TinyMCE‌ image edit dialog to have their
scale preselected. The precondition for this is that the old scales are
configured as picture variants.